Move GDAL helper functions into GDALUtils object
