+++ vlm/src/test/scala/geotrellis/contrib/vlm/RasterSummarySpec.scala
@@ -1,7 +1,5 @@
 package geotrellis.contrib.vlm
 
-import java.lang.management.ManagementFactory
-
 import geotrellis.contrib.vlm.gdal._
 import geotrellis.proj4._
 import geotrellis.raster._
@@ -10,8 +8,10 @@ import geotrellis.spark._
 import geotrellis.spark.testkit._
 import geotrellis.spark.tiling._
 import geotrellis.vector.Extent
+
 import spire.syntax.cfor._
 import org.apache.spark.rdd._
+
 import org.scalatest._
 
 class RasterSummarySpec extends FunSpec with TestEnvironment with BetterRasterMatchers with GivenWhenThen {
